requirements.txt
# REQUIREMENTS.TXT COMPLET POUR EDUCATIONAL AI PLATFORM
# =======================================================

# Core AI/ML dependencies
openai>=1.0.0
langchain>=0.1.0
langgraph>=0.1.0

# LangChain ecosystem - packages manquants
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-core>=0.1.0

# Pinecone (package renommé)
pinecone>=3.0.0

# Web interface
streamlit>=1.28.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.0.0

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
docx2txt>=0.8
tiktoken>=0.5.0

# Utilities
python-dotenv>=1.0.0
requests>=2.28.0

# Development & testing
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

# Optional: Security
cryptography>=41.0.0

# Database - SQLite
sqlalchemy>=2.0.0

# API Integration
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
requests>=2.28.0